rag-chatbot-backend/requirements.txt
fastapi
uvicorn
pydantic
groq
google-genai
qdrant-client==1.15.0
python-dotenv